Fixed nuke and react commands. Mod log doesn't ignore react messages anymore.

+++ src/commands/util/nuke.js
@@ -1,3 +1,7 @@
+'use strict';
+
+const { isUrl } = require('../../modules/type-checks');
+const { stripIndents } = require('common-tags');
 const Commando = require('discord.js-commando');
 const sendError = require('../../modules/send-error');
 
@@ -8,18 +12,23 @@ module.exports = class NukeCommand extends Commando.Command {
 			aliases: ['purge', 'prune', 'clean'],
 			group: 'util',
 			memberName: 'nuke',
-			description: 'Deletes messages.',
-			details: `Deletes messages. Here is a list of filters:
-				__invites:__ Messages containing an invite
-				__user @user:__ Messages sent by @user
-				__bots:__ Messages sent by bots
-				__uploads:__ Messages containing an attachment
-				__links:__ Messages containing a link`,
+			description: 'Delete messages. Filters available: user @User, me, bots, links, uploads',
+			details: stripIndents`Delete messages. Filters available:
+				**user @User:** Messages sent by @User
+				**me:** Messages sent by you
+				**bots:** Messages sent by bots
+				**links:** Messages containing a link
+				**uploads:** Messages containing an attachment`,
+			examples: [
+				'nuke 10',
+				'nuke 2 user @Kyuu#9348',
+				'nuke 5 links'
+			],
 			guildOnly: true,
 
 			args: [
 				{
-					key: 'limit',
+					key: 'numToDelete',
 					prompt: 'How many messages would you like to delete?',
 					type: 'integer',
 					max: 100
@@ -40,49 +49,61 @@ module.exports = class NukeCommand extends Commando.Command {
 		});
 	}
 
-	// TODO: fix filters lol
 	async run(msg, args) {
-		const limit = args.limit;
-		const filter = args.filter.toLowerCase();
-		let messageFilter;
+		let messages;
 
-		switch(filter) {
-		case 'bots':
-			messageFilter = message => message.author.bot;
-			break;
-		case 'links':
-			messageFilter = message => message.content.search(/https?:\/\/[^ \/\.]+\.[^ \/\.]+/) !== -1; // eslint-disable-line
-			break;
-		case 'invite':
-			messageFilter = (message =>
-					message.content.search(/(discord\.gg\/.+|discordapp\.com\/invite\/.+)/i) !== -1
-				);
-			break;
-		case 'upload':
-			messageFilter = message => message.attachments.size !== 0;
-			break;
-		case 'user':
-			if(!args.member) {
-				sendError(`Please provide a user (@User) to filter.`);
+		// Add 1 to the number to delete to account for the nuke message itself when deleting without a filter,
+		// deleting self, or deleting user that is self (for those who fail to use the correct filter).
+		const addCommandMsg = (!args.filter || args.filter === 'me' || msg.member.id === args.member.id);
+		if(addCommandMsg) args.numToDelete++;
+
+		if(!args.filter) {
+			messages = await msg.channel.fetchMessages({ limit: args.numToDelete });
+			await msg.channel.bulkDelete(messages.array().reverse());
+		} else {
+			try {
+				messages = await msg.channel.fetchMessages({ limit: 100 });
+				let messagesToDelete = messages.filterArray(this._getMessageFilter(msg, args));
+				if(messagesToDelete.length > args.numToDelete) {
+					messagesToDelete = messagesToDelete.slice(0, args.numToDelete);
+				}
+
+				if(messagesToDelete.length === 1) await messagesToDelete[0].delete();
+				else await msg.channel.bulkDelete(messagesToDelete.reverse());
+			} catch(err) {
+				return sendError(err);
 			}
-			messageFilter = message => message.author.id === args.member.user.id;
-			break;
-		case 'you':
-			messageFilter = message => message.author.id === message.client.user.id;
-			break;
 		}
 
-		msg.delete();
+		// Message was already deleted via mass delete, so don't try to delete it again.
+		if(addCommandMsg) return; // eslint-disable-line consistent-return
 
-		if(!filter) {
-			const messagesToDelete = await msg.channel.fetchMessages({ limit: limit });
+		return msg.delete();
+	}
 
-			msg.channel.bulkDelete(messagesToDelete.array().reverse());
-		} else {
-			const messages = await msg.channel.fetchMessages({ limit: limit });
-			const messagesToDelete = messages.filter(messageFilter);
+	_getMessageFilter(msg, args) {
+		const filter = args.filter.toLowerCase();
 
-			msg.channel.bulkDelete(messagesToDelete.array().reverse());
+		switch(filter) {
+		case 'user':
+			if(!args.member) {
+				throw new Error(`Please provide a user (@User) to filter.`);
+			}
+
+			return message => message.author.id === args.member.id;
+		case 'me':
+			return message => message.author.id === msg.member.id;
+		case 'bot':
+		case 'bots':
+			return message => message.author.bot;
+		case 'link':
+		case 'links':
+			return message => isUrl(message.content);
+		case 'upload':
+		case 'uploads':
+			return message => message.attachments.size !== 0;
+		default:
+			return null;
 		}
 	}
 };
